Rename azure_keyvault package
Consistent with JS and Python SDKs
