+++ sdk/keyvault/Cargo.toml
@@ -1,32 +1,37 @@
 [package]
-name = "azure_key_vault"
+name = "azure_keyvault"
 version = "0.1.0"
 authors = ["Microsoft Corp."]
 description = "Rust wrapper around Microsoft Azure REST APIs for Azure Key Vault"
 repository = "https://github.com/azure/azure-sdk-for-rust"
 homepage = "https://github.com/azure/azure-sdk-for-rust"
-documentation = "https://docs.rs/azure_key_vault"
+documentation = "https://docs.rs/azure_keyvault"
 keywords = ["azure", "microsoft", "sdk", "keyvault", "cloud"]
 categories = ["api-bindings"]
 readme = "README.md"
 license = "MIT"
 edition = "2018"
 
+[features]
+default = ["keys", "secrets"]
+keys = []
+secrets = []
+
 [dependencies]
-base64 = "0.13"
-thiserror = "1.0"
-reqwest = { version = "0.11", features = ["blocking", "json"] }
-chrono = { version = "0.4", features = ["serde"] }
-const_format = "0.2.13"
-serde_json = "1.0"
-url = "2.2"
-serde = { version = "1.0", features = ["derive"] }
-getset = "0.1"
-oauth2 = "4.0.0"
 azure_core = { path = "../core", version = "0.1.0" }
 azure_identity = { version = "0.1", path = "../identity" }
+base64 = "0.13"
+chrono = { version = "0.4", features = ["serde"] }
+const_format = "0.2.13"
+getset = "0.1"
+oauth2 = "4.0.0"
+reqwest = { version = "0.11", features = ["blocking", "json"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+thiserror = "1.0"
+url = "2.2"
 
 [dev-dependencies]
-mockito = "0.30"
 async-trait = "0.1"
+mockito = "0.30"
 tokio = { version = "1.0", features = ["full"] }
