+++ sdk/keyvault/examples/delete_secret.rs
@@ -1,5 +1,5 @@
 use azure_identity::token_credentials::{ClientSecretCredential, TokenCredentialOptions};
-use azure_key_vault::KeyClient;
+use azure_keyvault::KeyClient;
 use std::env;
 
 #[tokio::main]
@@ -10,7 +10,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let tenant_id = env::var("TENANT_ID").expect("Missing TENANT_ID environment variable.");
     let keyvault_url =
         env::var("KEYVAULT_URL").expect("Missing KEYVAULT_URL environment variable.");
-    let backup_blob = env::var("BACKUP_BLOB").expect("Missing BACKUP_BLOB environment variable.");
+    let secret_name = env::var("SECRET_NAME").expect("Missing SECRET_NAME environment variable.");
 
     let creds = ClientSecretCredential::new(
         tenant_id,
@@ -19,8 +19,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         TokenCredentialOptions::default(),
     );
     let mut client = KeyClient::new(&keyvault_url, &creds)?;
-
-    client.restore_secret(&backup_blob).await?;
+    client.delete_secret(&secret_name).await?;
 
     Ok(())
 }
