+++ sdk/keyvault/examples/set_secret.rs
@@ -1,5 +1,5 @@
 use azure_identity::token_credentials::{ClientSecretCredential, TokenCredentialOptions};
-use azure_key_vault::KeyClient;
+use azure_keyvault::KeyClient;
 use std::env;
 
 #[tokio::main]
@@ -11,6 +11,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let keyvault_url =
         env::var("KEYVAULT_URL").expect("Missing KEYVAULT_URL environment variable.");
     let secret_name = env::var("SECRET_NAME").expect("Missing SECRET_NAME environment variable.");
+    let secret_value =
+        env::var("SECRET_VALUE").expect("Missing SECRET_VALUE environment variable.");
 
     let creds = ClientSecretCredential::new(
         tenant_id,
@@ -19,7 +21,11 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         TokenCredentialOptions::default(),
     );
     let mut client = KeyClient::new(&keyvault_url, &creds)?;
-    client.delete_secret(&secret_name).await?;
+
+    client.set_secret(&secret_name, &secret_value).await?;
+
+    let secret = client.get_secret(&secret_name).await?;
+    assert_eq!(secret.value(), "whatup");
 
     Ok(())
 }
