+++ sdk/keyvault/src/client.rs
